+++ lib/genPouch.js
@@ -74,7 +74,7 @@ async function docCount(database) {
     try {
       remoteDocCount= await db.info()
     } catch (err) {
-      remoteDocCount = await db.info();
+      remoteDocCount = await db.info(); // TODO Why is this like that?
       console.log(err.message);
     } finally {
       localDocCount = await db.info()
