+++ lib/monitoring.js
@@ -1,0 +1,6 @@
+const monitor = require('os-monitor')
+let res = {}
+
+monitor.start({ stream: true }).pipe(res)
+
+module.exports = res
